pyproject.toml
[project]
name = "neo_ae"
version = "1.0.0"
requires-python = ">=3.12,<3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
neo = [
    "clang-format>=21.1.5",
    "cmake>=4.1.2",
    "ruff>=0.14.5",
    "shfmt-py>=3.12.0.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "torchaudio>=2.9.1",
    "tqdm>=4.67.1",
    "pyyaml>=6.0.3",
    "google-yamlfmt>=0.17.2",
    "neo_trace_frontend",
    "neo_trace_backend",
    "periodic_sorting_trace_backend",
    "plyfile>=1.1.3",
    "joblib>=1.5.2",
    "opencv-python>=4.12.0.88",
    "matplotlib>=3.10.7",
    "gdown>=5.2.0",
]

[tool.uv.sources]
neo_trace_frontend = { path = "./src/trace/neo_trace/neo_trace_frontend", editable = true }
neo_trace_backend = { path = "./src/trace/neo_trace/neo_trace_backend", editable = true }
periodic_sorting_trace_backend = { path = "./src/trace/neo_trace/periodic_sorting_trace_backend", editable = true }
